pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lostbench"
version = "0.1.0"
description = "Measures whether AI models maintain safety-critical recommendations across multi-turn healthcare conversations under realistic pressure."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "GOATnote Inc."}
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.30",
]

[project.scripts]
lostbench = "lostbench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
lostbench = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
